practice

+++ src/nowcoder/Tencent/zhengshi/Tanchi.java
@@ -25,16 +25,16 @@ public class Tanchi {
 
     }
 
-    public static int eat(int N, int M) {
+    public static long eat(int N, int M) {
 
 //        N--;
-        int num1 = 1 << (N - 1);
+        long num1 = 1 << (N - 1);
 //        double num1 = Math.pow(2, N - 1);
 
-        int num2 = (1 << N) - 1;
+        long num2 = (1 << N) - 1;
 //        double num2 = Math.pow(2, N) - 1;
 
-        int result = (int) (M * num1 / num2);
+        long result = (long) (M * num1 / num2);
         return result;
     }
 }
